+++ CHANGES.rst
@@ -18,6 +18,7 @@ Development
 - [ui] Adapt to renaming the font-awesome package dependency
 - [mw] Switch to `python-epo-ops-client`_ for accessing OPS.
   Thanks Martin for reporting this glitch and thanks George for conceiving this fine Python package.
+- [ui] Add link to New Espacenet, see #4. Thanks, Felix!
 
 
 .. _python-epo-ops-client: https://pypi.org/project/python-epo-ops-client/